requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
django==4.2.24
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
